move everything into fields module
